JARVIS2/requirements.txt
SpeechRecognition==3.10.4
PyAudio==0.2.14
pyttsx3==2.90
requests==2.32.3
python-dotenv==1.0.1
pyautogui==0.9.54

